Spark 1735 small UI improvements
* Line-ending madness

I've given up on trying to avoid having to make this commit. Somehow, git
insists on changing these files.

* SPARK-1735: Improvements on resizing the MUC window.

The (right-hand) list of users should not be sized relative to the overal size
of the window. Instead, it should be just large enough for the names to fit. It
also does not need to shrink or grow when the window is resized.

* SPARK-1735: Removing redundant call.

* SPARK-1735: When using bookmarks, ensure room name is used.

MUC Bookmarks set by some clients include the room description as its name. When
Spark automatically joins a room based on a bookmark, it should not use that
description as the room name. Instead, ensure that the node-part of the JID is
used.

* SPARK-1735: Fix presence in MUC.

Presence information in MUC context should be obtained from the room, not
from the roster of the Spark user.