SPARK-1741: StringBuilder -> String
Reports any variables declared as or uses of java.lang.StringBuffer and java.lang.StringBuilder which can be
replaced with a single java.lang.String concatenation. Using a String concatenation makes the code shorter and
simpler. This inspection only reports when the resulting concatenation is at least as efficient or more
efficient than the original StringBuffer or StringBuilder use.