SPARK-1740: Allow SASL GSS-API to be executed as it was in SMACK 3
SMACK 4 uses the fully qualified host name, when negotiating GSS-API
via SASL. SMACK 3 used the XMPP domain name instead. Spark should
have a configurable option that allows it to fall back to the old
mechanism (which is likely needed to connect to servers that
have already been configured to accept SSO from older versions of
Spark).

This commit also brings a reorganization of the code of the
'advanced connection settings' dialog - simply because it was pretty
unpractical as it was. Small modifications have been made to the
layout.