Spark 1740 smack3 compat (#200)
* SPARK-1735: MUC occupants list should have width

This commit attempts to prevent that the occupants list in a MUC
screen has a near-zero width.

* SPARK-1772: Icons for muc role fix

* SPARK-1735: MUC occupants list should have vertical scrollbar

* SPARK-1740: Allow SASL GSS-API to be executed as it was in SMACK 3

SMACK 4 uses the fully qualified host name, when negotiating GSS-API
via SASL. SMACK 3 used the XMPP domain name instead. Spark should
have a configurable option that allows it to fall back to the old
mechanism (which is likely needed to connect to servers that
have already been configured to accept SSO from older versions of
Spark).

This commit also brings a reorganization of the code of the
'advanced connection settings' dialog - simply because it was pretty
unpractical as it was. Small modifications have been made to the
layout.

* SPARK-1740: Allow SASL GSS-API to be executed as it was in SMACK 3