SPARK-1791: Distributales should include plugins.
The earlier structure generated the Spark distributables from the 'core' module. This module did not contain optional features (notably: plugins). This in turn made it impossible to include those optional features in the distributables.

To work around this problem, the distributables are now created in a dedicated module (aptly named 'distribution'). This module defines dependencies and assembly references to other modules of which artifacts need to be included in the distributables.

As a proof-of-concept, the 'roar' plugin was re-instated.