SPARK-1791: Replaced lobobrowser with JxBrowser (eval)
The project for the embedded browser that we used, lobobrowser, is defunct. Maven artifacts are not to be found.

A possible replacement is JxBrowser (by TeamDev). This software is proprietary, but they do offer free licenses to open source software.

In order to evaluate wether we'd like to apply for one, this commit replaces lobobrowser with JxBrowser. An evaluation license is included, but will no longer be valid in 30 days from now.