SPARK-1791: Restoring the internal 'idle' plugin.
The 'idle' plugin was an internal (in core) that was initially not migrated from Ant to Maven. It depended on both an unknown libary, as well as the 'phone' internal plugin, which had issues of its own.

This commit restores the 'idle' plugin, by providing a reference to the previously unknown library (JNA), and removing the dependency to the 'phone' internal plugin.