SPARK-1791: Populate plugin.xml from pom.xml
Both the pom.xml file and the plugin.xml file define the plugin name, version and description. It is error prone to keep both files in sync. Instead, the plugin.xml file should be popluated from the data that is in the pom.xml file.

Maven resource filtering is employed for this purpose. The drawback of this approach is that the plugin.xml file needs to be a Maven resource, which makes it part of the jar file. This means that this file is included twice in each plugin: once in the 'outer' plugin jar file, and once in the jar file that contains the plugin's compiled sources.