SPARK-1757 & SPARK-1649: Don't process MUC history as regular messages.
when joining a multi user chat room, the recent history of that room can be sent to the client.

A bug in Spark caused these mesages to be processed as history (which is good), but also as
regular/new messages (which is wrong). As a result of this, the messages could be presented in
the wrong order on screen (SPARK-1757), as the original date of the message was not processed.
Also, all event listeners got triggered, causing other components to act as if a new message
was received. This causes issues like SPARK-1649 (unwanted notifications). Finally, the history
was being saved in the local history - which can introduce some duplication there (when a room
is rejoined frequently).