SPARK-1812: Parse vCards on disk
This commit revisits an earlier fix for SPARK-1812. Instead of skipping fixed number of XML nodes,
the parser should skip until at the start of the vCard element that's identified by name/namepace.

Also, file references should be closed, even in case of an error (the try-with-resources construct
will make sure this happens).

Finally, when a file cannot be parsed, that file should be deleted.