SPARK-945: Save component bounds upon change
This commit changes how component bounds (position + dimension) are stored. Before this commit, this
occurred only when every component was closing. When closing Spark abruptly, these event listeners did not always trigger.

With this commit, the bounds of components are updated whenever they change. The bounds are saved just once, when the main screen quits.