SPARK-1962: Don't force smilies to lowercase.
The old code was doing "smiley-text to image" lookup by converting everything to lowercase, then doing a
case insensitive lookup. My last commit made that lookup case sensitive, which caused lookups for :-D and :-P to fail.

This commit remove the to-lowercase conversion of the text that's used in the lookup, fixing those instances again.