SPARK-2029: Don't check for voice
Upon receiving messages to be displayed in a chat room UI, don't explicity check if the
sender has voice (the server will have done this). This commit replaces the check with
a check to verify that the sender is a participant in the room (and I'm not even sure if
that's warranted).