Fix Spark reconnection
Smack's ReconnectionManager was running and doing its job, but since
the ReconnectionListener API has changed, especially the
reconnectionSuccessful() callback was removed, Spark's GUI components
where not aware that the reconnection was successful. The
reconnectionSuccessful() callback got remove because it is identical
with then authenticated() callback from the ConnectionListener. Thus
this commit basically replaces reconnectionSuccessful() with
authenticated().