SPARK-2133: Don't search jar for translations
Instead of specifically looking for i18n-files in the spark.jar file, just check if they're anywhere on the classpath. This is more likely to yield results (especially when Spark isn't ran from a jar file, but from an IDE). I'd not be surprised if it's moderately faster too, as no ZIP file operations are needed).