SPARK-2079: Ensure settings are properly saved upon login
The way the 'advanced settings' dialog of the login screen works, causes it to save its configuration to file whenever 'an event' happens.

When saving the configuration, the state of the various UI elements (checkboxes, radiobuttons, etc) is used to update the configuration file.

Eligble events happen more often than expected, for example, when the dialog is disposed (which can occur much later after when it was closed). By that time, the settings of the UI elements can be outdated. This causes an issue where settings are unexpectedly 'reverted'.

This commit works around this problem by only allowing the settings to be updated while the screen is visible.