SPARK-2138: TicTacToe - Replace AsyncStanzaListener with IqRequestHandler
When processing an incoming IQ request (such as a game invitation), the version of Smack that we use does not allow us to use an AsyncStanzaListener anymore.

Before all registered stanza listeners are even invoked, Smack has determined that no IqRequestHandler has been registered to process the request, and will already have sent back an error.