Increase version number of all plugins
Since the last release, major changes were applied to the project. To reflect this, I've increased the version number of all of the plugins.

Not all plugins might have had functional changes, but given that the project structure has changed a lot, some kind of increase (that will allow us to tell plugins before and after this release apart) could be helpful.