SPARK-2154: Do not parse nicknames using Jid implementation
Nicknames aren't necessarily components of a Jid (they are in context of MUC, but they're pretty much free form text on the roster). By attempting to use the Jid implementation to parse nicknames, exceptions occur. In this instance, such an exception prevents an offline message being displayed if it is sent by someone that is on the recipients roster using in nickname that contains a space.