SPARK-2184: Fix the validation of Basic Constraints in the cert chain
The old code seems to have this backwards. All certificates in the chain, except for the leaf (end-entity) should have Basic Constraints.

When validating a certificate chain, Spark should make sure that the pathLen attribute of the Basic Constraint extension is valid and the cA field is set to TRUE for each non-leaf certificate.