SPARK-2194: Fix certificate expiry configuration
The code to skip the date validation is using an API method that is intended for a very similar, but different purpose.

The change in this commit replaces it, by allowing for the validation to throw an exception, but only re-throw that exception if Spark is not configured to suppress it.