SPARK-2360: Isolate bugs, log improvements
There is a slew of minor issues, such as operating on non-existing images, that throw unchecked exceptions and break the application badly.

To isolate these occurrences, exception handling can be put in place, with some logging being generated. That won’t solve the underlying issue, but it should limit the impact of these issues.